+++ miniapp/src/main/java/com/rakuten/tech/mobile/miniapp/MiniAppSdkConfig.kt
@@ -1,5 +1,7 @@
 package com.rakuten.tech.mobile.miniapp
 
+import android.util.Log
+
 /**
  * This represents the configuration settings for the Mini App SDK.
  * @property baseUrl Base URL used for retrieving a Mini App.
@@ -38,6 +40,8 @@ data class MiniAppSdkConfig(
     )
 
     init {
+        Log.d("AAAAA",""+isPreviewMode)
+
         when {
             !isBaseUrlValid(baseUrl) ->
                 throw sdkExceptionForInvalidArguments("MiniAppSdkConfig with invalid baseUrl")